chore(web): List layout & list refactor
